Identify cable and flett cards
Some pcieslots support both cable and flett cards. To expose
the ports corresponding to the card a list of CCINs is added
to each port's inventory path in the VPD json files. Merging
this commit will expose ports belongs to the card.

Testing:
Verified the ports corresponding to the installed card are exposed.
Here is an example output from
'busctl tree xyz.openbmc_project.Inventory.Manager':
/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10
    └─/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10/pcie_card10
    ├─/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10/pcie_card10/c10_connector1
    ├─/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10/pcie_card10/c10_connector2
    ├─/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10/pcie_card10/c10_connector3
    └─/xyz/openbmc_project/inventory/system/chassis/motherboard/pcieslot10/pcie_card10/c10_connector4

Signed-off-by: Shantappa Teekappanavar <sbteeks@yahoo.com>