Refresh VINI/CC for PCIe while priming inventory
As the VPD data is persisted in BMC, VPD can be present for
PCIe devices which is not present.
Entity manager probes VINI/CC to check for PCIe device which
leads to ambiguity.

This commit clears VINI/CC at the time of priming inventory
path for PCIe devices in case it is absent.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>