Update backup and restore status
The commit adds a property to hold status of backup and restore
operation. The information is required to detect any error in the
process of performing backup and restore or current status of the
operation all together.

As per current implementation the functionality could not get
re-triggered once it has been executed. The backupAndRestore API
keeps a check against this restriction using the status property
added via this commit.

In case user wants to overwrite this and trigger the functionality
multiple times, user can overwrite the status property before
calling the functionality. To achieve this a setter method has been
provided as well.