Default keyword value update for backup & restore
Default keyword values from the backup and restore config JSON file
was assumed that, it will be considered as hex value after parsing
the config JSON file in the code.
But it was directly considered as decimal value instead as hex value
after parsing JSON file and these default values were used for
comparing against source and destination’s keywords hex values.
This was causing invalid comparison.

To avoid this issue config JSON file is updated with proper default
decimal values, which is the correct counter value of the hexadecimal
needed for comparison with source and destination keyword value,
in this commit.