Instantiate event handler to poll on FRU presence
GPIO pins helps in detecting the presence of the FRU via presence pins.
Monitoring these pins helps in detecting the FRUs if they are removed
or added.

This commit adds the code to collect the GPIO pins data from config JSON
that requires polling and instantiates event handler object for the pins
to do event handling when there is any change in the pin presence.

Signed-off-by: RekhaAparna01 <vrekhaaparna@ibm.com>