Fix: vpd-tool input validation checks
vpd-tool with option’s which does not require an object path is
failing with an error ‘Given path is empty’.

As CLI library checks whether value is passed or not for the
selected options by default. If the value is passed as empty string
as “” is not being caught by CLI library.

As CLI library checks whether value is passed or not for the selected
options by default. If the value is passed as empty string as “” is not
being caught by CLI library. Added additional validation that passed
object path, record names are not empty.