Moving out updateKeyword core logic from manager
This commit adds the code to move updateKeyword API from
Manager class to Parser class. And removed the dependent API
from the Manager class.

As updateKeyword API implementation core logic
is required in the backup&restore functionality to update
the keyword’s value. So, Parser class will host this functionality.

In the Parser class updateVpdKeyword API is introduced, which updates
keyword’s value on the given EEPROM path and updates on DBus,
redundant EEPROM path(s) also if those path’s information is available
in the system config JSON.