Post action tag in JSON
The commit introduces an action tag "postAction".

This can be used in case a set of job is required to be executed
after the collection has been done.

Post action can also be tagged to be performed for some specifc
FRUs. For this CCIN list can be added under it and based on that
it will be executed.
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>