Clear vpd before recollection
Some FRUs qualifies for recollection at standby.
To avoid any misleading representation because of the stale data,
VPD is being cleared and present is set to false before
recollection is triggered for those FRUs.
The commit also checks for any sub tree under the parent path and
if found resets data for them as well.

This will also help in detecting any error that can happen during
recollection of those FRUs as their present property will remain
set to false.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>