Update present property for subtree
When BMC is rebooted at chassis off state, all the FRUs are queued
for recollection and their present property is set accordingly.
Whereas in case there is any sub fFRU under that parent, which was
not populated based on udev event then its present property remains
what was there before boot, creating a wrong representatin of the
system.

The commit fixes present for those sub FRUs.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>