Everest:Add Slot interface to all NVMe drives
Assign a slot number to all drive paths so that the entity manager
config file for IBM's drives can use that to give its objects unique
names.  Previously it used its own incrementing index but sometimes got
confused and tried to put a duplicate interface on the same object path
which would crash.

This was already being done for IBM's cable card sensors.

Because the NVMe temperature sensor names are being hardcoded in other
places (fan control and HMC telemetry), the slots are being assigned to
always start at 1 and sequentially increment to keep the names the same
as before.

Tested:

Properties are on D-Bus:

```
 busctl get-property xyz.openbmc_project.Inventory.Manager \
 /xyz/openbmc_project/inventory/system/chassis/motherboard/disk_backplane0/nvme0/dp0_drive0 \
 xyz.openbmc_project.Inventory.Decorator.Slot SlotNumber
u 1
```

Change-Id: I233397450dc47ccc8c02a01bba1fa42e29db857c
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>