Backup restore JSON for Everest system
This commit adds backup and restore config JSON file for Everest
system.

Tested the changes by loading image on the Everest system,
* Manually set the VSYS BR record value on DBus using busctl to create
mismatch value with the value present on hardware, observed that the
mismatch value is retained as it is on DBus and on the hardware
after restarting vpd-manager application.
*  Set the VSYS BR record value on DBus with its default value using
busctl command, after vpd-manager application restart BR value from the
hardware is reflected on the DBus.

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>