Update one time properties
The commit populates and initializes "Functional" and "Enabled"
property for a given FRU.

The properties are only populated and initilized during first
collection of the FRU. As the value of these properties are not
controlled by VPD-Collection process, they are not re-populated
in case they are found already existing over DBus.
This is done to retain the property set for these properties
across BMC reboots or any re-collection.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>