Fix: Pretty Name not showing on bmc UI
For the defect 665516, where name of VRM and fabric adapters in
inventory page is not coming correctly.

In insertOrMerge API insert call on std::map object was not overwriting
properties value, if the property key is already found in the map.
Which causing empty PrettyName value got published on the DBus. Hence
bmc web was not showing correct prettyName for FRUs, who’s VPD
collection is failed and prime inventory is called.

This commit has fix for the above issue. I tested the changes, observed
name was showing correctly for VRM and fabric adapters on bmc web.

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>