Skip PCIe cards collection at power on
Collection of PCIe card is skipped in case of BMC reboot at chassis
power on.

Since PCIe cards are CMable and not hot pluggable, any removal or
addition of card will be handled in the CM flow. No need to explicit
collection when BMC is coming up while chassis is on.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>