JSON utility error code handling -1 (#662)
* Update json_utility APIs with error codes

This commit updates API getVPDOffset and getInventoryObjPathFromJson to
set error codes in case of errors/exceptions. This helps caller of the
APIs to take action based on the error code returned from the API.

Change-Id: Icfe2640ee311d61e83ca302e5199ce4dd1057c8b
Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>

* Set error code for getFrusWithPresenceMonitoring

This commit updates getFrusWithPresenceMonitoring API to set error code
in case of error. This helps caller of API to take action based on the
error code returned from the API.

Change-Id: I302dca2e7e667ebd087322761f7c180a21fb454e
Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>

* Set error code for isFruPresenceHandled API

This commit updates isFruPresenceHandled API to set error code in case
of error. This helps caller of API to take action based on the error
code returned from the API.

Change-Id: I3fe4c35edbeeb889c2160705b1b2aece12104657
Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>

* Set error code for getParsedJson API

This commit updates getParsedJson API to set error codes in case of
error. This helps caller of the API to take action based on the error
code returned from the API.

Change-Id: I0b88806bbc2201ae162e1714613fa6a9bfe3df9d
Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>

* Set error code for isFruReplaceableAtStandby API

This commit updates isFruReplaceableAtStandby API to set error code in
case of error. This helps caller of API to take action based on the
error code returned from the API.

Change-Id: I032ae984fefd9d1fc419c88a399db527f028f4fa
Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>

---------

Signed-off-by: Rekha Aparna <vrekhaaparna@ibm.com>
Co-authored-by: Rekha Aparna <vrekhaaparna@ibm.com>