Fix sanity check failure due to unwritten data (#681)
The 'write' keyword operation for module VPD was failing during the
sanity check with an ECC error. This occurred because written data was
not yet synced to the file when the sanity check was performed.

To address this, an explicit flush is now called after writing the data
to ensure it is written to the file immediately. With this change, the
ECC check no longer fails during the sanity check while updating VPD.

The commit also enables sanity check back.

Change-Id: If4fbda85310422f694e421018db5900b1bb5e6be

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>